updated image upload backend,successfully running, front-end should be checked

+++ app/src/main/java/com/project/app/repo/FormRepo.java
@@ -10,7 +10,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface FormRepo extends JpaRepository<Event,Integer> {
+public interface FormRepo extends JpaRepository<Event,Long> {
 
     
     List<Event> findByisPublicTrue();
